requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
deta==1.1.0
estuary-client==1.0.6
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
instantcli==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pestuary==1.0.14
pi==0.1.2
Pillow==9.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
qrcode==7.3.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
